requirements.txt
streamlit
streamlit-folium
folium
google-generativeai
Pillow
streamlit-geolocation
